feat: add remaining crud unit tests and adjusts some configs

+++ src/modules/customers/application/services/delete-customer.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CustomerRepository } from '../../domain/customer.repository';
 
 /**
@@ -10,6 +10,11 @@ export class DeleteCustomerService {
   constructor(private readonly customerRepository: CustomerRepository) {}
 
   async execute(id: number): Promise<void> {
+    const customer = await this.customerRepository.findById(id);
+    if (!customer || !customer.status) {
+      throw new NotFoundException(`Customer with ID ${id} not found`);
+    }
+
     await this.customerRepository.delete(id);
   }
 }
